requirements.txt
PyGithub
pandas
python-dotenv
numpy
fastapi[standard]
uvicorn
httpx
ipykernel
scikit-learn
google-generativeai
scikit-learn
joblib
